Some major changes added enemy to fire bullet
